+++ docs/changelog.md
@@ -5,6 +5,8 @@
 ### Bug fixes
 
 - fixed `run custom instruction` not working when program was loaded but not started
+- fixed exiting `debug select` mode not returning to default program state when started when the program was loaded but not started
+- fixed `debug select` keybinding hit not updating when `debug select` mode was exited when it was started when the program was loaded but not started 
 
 ## v1.8.0 (latest version)